move setup cc_net into setup.py
